requirements.txt
flaskext
flask
flask_cors
pymysql
mysql-connector-python